refactor mutations into their own types based on entity
